requirements.txt
anitopy
beautifulsoup4
fastapi
fastapi-cache2
httpx[http2]
loguru
lxml
uvicorn
